pyproject.toml
[tool.poetry]
name = "python_bot"
version = "0.1.0"
description = ""
authors = ["Evan Vilagov <evan.vilagov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pyTelegramBotAPI = "^3.7.7"
python-dotenv = "^0.17.0"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
